Correct a few analyzers in Calibration/HcalCalibAlgos aganist static analyzers
