+++ Calibration/HcalCalibAlgos/src/classes.h
@@ -1,4 +1,4 @@
-#include "Calibration/HcalCalibAlgos/src/TCell.h"
+#include "Calibration/HcalCalibAlgos/interface/TCell.h"
 
 namespace Calibration_HcalCalibAlgos {
   struct dictionary {};
